+++ includes/parser.h
@@ -6,7 +6,7 @@
 /*   By: takkatao <takkatao@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/23 16:11:50 by takkatao          #+#    #+#             */
-/*   Updated: 2022/04/23 16:11:50 by takkatao         ###   ########.fr       */
+/*   Updated: 2022/04/23 17:30:52 by takkatao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,5 +22,6 @@
 
 void		free_exec_info(t_exec_info *exec_info);
 t_exec_info	*parser(t_token *token);
+void		init_parser(t_exec_info **e, t_token *t, int *i, int *j);
 
 #endif
